move test_init_out_dir
